pc_stats_logs/ram_usages/requirements.txt
psutil
GPUtil
python-dotenv
SQLAlchemy[asyncio]
asyncpg
pydantic
